requirements.txt
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
factory-boy==2.12.0
drf_yasg==1.17.1
# prod
gunicorn==20.0.4
psycopg2==2.8.4
whitenoise==5.0.1
dj_database_url==0.5.0